refactor: RewriteRuleの参照を変更
