+++ host-frontend-root/frontend-src-root/src/application/ports/IRewriteRuleRepository.ts
@@ -1,4 +1,4 @@
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 import { RewriteRules } from 'src/domain/value-objects/RewriteRules';
 
 export interface IRewriteRuleRepository {
