+++ host-frontend-root/frontend-src-root/src/application/usecases/rule/LoadRewriteRuleForEditUseCase.ts
@@ -1,5 +1,5 @@
 import { IRewriteRuleRepository } from 'src/application/ports/IRewriteRuleRepository';
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 
 export class LoadRewriteRuleForEditUseCase {
   constructor(
