+++ host-frontend-root/frontend-src-root/src/components/organisms/RulesTable/RulesTable.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import RuleTableRow from 'src/components/molecules/RuleTableRow/RuleTableRow';
 import styles from 'src/components/organisms/RulesTable/RulesTable.module.css';
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 
 interface RulesTableProps {
   rules: RewriteRule[];
