+++ host-frontend-root/frontend-src-root/src/domain/entities/ReplaceElementPreservingState.ts
@@ -1,6 +1,6 @@
 import { RegexConstants } from 'src/domain/constants/RegexConstants';
 import { ParserContextStrategyFactory } from 'src/domain/entities/ParserContextStrategy';
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 import { IElementFactory } from 'src/domain/ports/IElementFactory';
 
 /**
