+++ host-frontend-root/frontend-src-root/src/domain/value-objects/Tab.ts
@@ -1,4 +1,4 @@
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 import { TabId } from 'src/domain/value-objects/TabId';
 import { TabUrl } from 'src/domain/value-objects/TabUrl';
