+++ host-frontend-root/frontend-src-root/src/domain/value-objects/Tabs.ts
@@ -1,4 +1,4 @@
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 import { Tab } from 'src/domain/value-objects/Tab';
 
 /**
