+++ host-frontend-root/frontend-src-root/src/infrastructure/browser/messaging/ChromeRuntimeRewriteRuleRepository.ts
@@ -1,5 +1,5 @@
 import { IRewriteRuleRepository } from 'src/application/ports/IRewriteRuleRepository';
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 import { RewriteRuleNotFoundError } from 'src/domain/errors/RewriteRuleNotFoundError';
 import { RewriteRules } from 'src/domain/value-objects/RewriteRules';
