+++ host-frontend-root/frontend-src-root/tests/unit/domain/entities/ElementMatchesFlexiblePattern/exec/regex-pattern-matching.test.ts
@@ -1,7 +1,7 @@
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
 import { ElementMatchesFlexiblePattern } from 'src/domain/entities/ElementMatchesFlexiblePattern';
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 
 describe('ElementMatchesFlexiblePattern.exec() - Regex Pattern Matching (isRegex: true)', () => {
   let container: HTMLElement;
