+++ host-frontend-root/frontend-src-root/tests/unit/domain/entities/RewriteRule/createRedundantPattern/regex-pattern.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect,it } from 'vitest';
 
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 
 /**
  * RewriteRule.createRedundantPattern 正規表現パターンテストケース
