+++ host-frontend-root/frontend-src-root/tests/unit/domain/entities/RewriteRule/fromPlainObject/error-cases.test.ts
@@ -5,7 +5,7 @@
  */
 import { describe, expect,it } from 'vitest';
 
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 
 describe('RewriteRule.fromPlainObject - 異常系', () => {
   it('should handle partially missing required parameters', () => {
