+++ host-frontend-root/frontend-src-root/tests/unit/domain/entities/RewriteRule/fromPlainObject/normal-cases.test.ts
@@ -8,7 +8,7 @@
  */
 import { describe, expect,it } from 'vitest';
 
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 
 describe('RewriteRule.fromPlainObject - 正常系', () => {
   it('should create RewriteRule instance from plain object with all parameters', () => {
