+++ host-frontend-root/frontend-src-root/tests/unit/domain/entities/RewriteRule/fromPlainObject/validation-error-cases.test.ts
@@ -4,7 +4,7 @@
  */
 import { describe, expect,it } from 'vitest';
 
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 
 describe('RewriteRule.fromPlainObject - バリデーションエラー', () => {
   const invalidInputs = [
