+++ host-frontend-root/frontend-src-root/tests/unit/domain/entities/RewriteRule/matchesUrl/normal-cases-false.test.ts
@@ -7,7 +7,7 @@
  */
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 
 describe('RewriteRule.matchesUrl - 正常系 (falseを返すケース)', () => {
   beforeEach(() => {
