+++ host-frontend-root/frontend-src-root/tests/unit/domain/value-objects/RewriteRules/applyRulesWithDomDiffer/normal-cases.test.ts
@@ -5,7 +5,7 @@
 import { createMockElementFactory } from 'tests/unit/domain/ports/IElementFactory/createMockElementFactory';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 import { RewriteRules } from 'src/domain/value-objects/RewriteRules';
 
 const mockElementFactory = createMockElementFactory();
