+++ host-frontend-root/frontend-src-root/tests/unit/domain/value-objects/RewriteRules/constructor/normal-cases.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach,describe, expect, it } from 'vitest';
 
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 import { RewriteRules } from 'src/domain/value-objects/RewriteRules';
 
 /**
