+++ host-frontend-root/frontend-src-root/tests/unit/infrastructure/persistence/indexeddb/DexieRewriteRuleRepository/create/Abend/error-cases.test.ts
@@ -2,7 +2,7 @@ import '../../../setup';
 
 import { afterEach,beforeEach, describe, expect, it } from 'vitest';
 
-import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
+import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 import { dexieDatabase } from 'src/infrastructure/persistence/indexeddb/DexieDatabase';
 import { DexieRewriteRuleRepository } from 'src/infrastructure/persistence/indexeddb/DexieRewriteRuleRepository';
